+++ src/aggregators/ccnrm.jl
@@ -3,8 +3,8 @@
 # algorithm with optimal binning,  Journal of Chemical Physics 143, 074108
 # (2015). doi: 10.1063/1.4928635.
 
-mutable struct CCNRMJumpAggregation{T, S, F1, F2, RNG, DEPGR, PT} <:
-               AbstractSSAJumpAggregator{T, S, F1, F2, RNG}
+mutable struct CCNRMJumpAggregation{T, S, F1, F2, DEPGR, PT} <:
+               AbstractSSAJumpAggregator{T, S, F1, F2}
     next_jump::Int
     prev_jump::Int
     next_jump_time::T
@@ -15,15 +15,14 @@ mutable struct CCNRMJumpAggregation{T, S, F1, F2, RNG, DEPGR, PT} <:
     rates::F1
     affects!::F2
     save_positions::Tuple{Bool, Bool}
-    rng::RNG
     dep_gr::DEPGR
     ptt::PT
 end
 
 function CCNRMJumpAggregation(nj::Int, njt::T, et::T, crs::Vector{T}, sr::T,
-        maj::S, rs::F1, affs!::F2, sps::Tuple{Bool, Bool},
-        rng::RNG; num_specs, dep_graph = nothing,
-        kwargs...) where {T, S, F1, F2, RNG}
+        maj::S, rs::F1, affs!::F2, sps::Tuple{Bool, Bool};
+        num_specs, dep_graph = nothing,
+        kwargs...) where {T, S, F1, F2}
 
     # a dependency graph is needed and must be provided if there are constant rate jumps
     if dep_graph === nothing
@@ -45,30 +44,31 @@ function CCNRMJumpAggregation(nj::Int, njt::T, et::T, crs::Vector{T}, sr::T,
         binwidthconst = binwidthconst, numbinsconst = numbinsconst) # We will re-initialize this in initialize!()
 
     affecttype = F2 <: Tuple ? F2 : Any
-    CCNRMJumpAggregation{T, S, F1, affecttype, RNG, typeof(dg), typeof(ptt)}(
+    CCNRMJumpAggregation{T, S, F1, affecttype, typeof(dg), typeof(ptt)}(
         nj, nj, njt, et,
         crs, sr, maj,
         rs, affs!, sps,
-        rng, dg, ptt)
+        dg, ptt)
 end
 
-+############################# Required Functions ##############################
+############################# Required Functions ##############################
 # creating the JumpAggregation structure (function wrapper-based constant jumps)
 function aggregate(aggregator::CCNRM, u, p, t, end_time, constant_jumps,
-        ma_jumps, save_positions, rng; kwargs...)
+        ma_jumps, save_positions; kwargs...)
 
     # handle constant jumps using function wrappers
     rates, affects! = get_jump_info_fwrappers(u, p, t, constant_jumps)
 
     build_jump_aggregation(CCNRMJumpAggregation, u, p, t, end_time, ma_jumps,
-        rates, affects!, save_positions, rng; num_specs = length(u),
+        rates, affects!, save_positions; num_specs = length(u),
         kwargs...)
 end
 
 # set up a new simulation and calculate the first jump / jump time
 function initialize!(p::CCNRMJumpAggregation, integrator, u, params, t)
     p.end_time = integrator.sol.prob.tspan[2]
-    initialize_rates_and_times!(p, u, params, t)
+    rng = get_rng(integrator)
+    initialize_rates_and_times!(p, u, params, t, rng)
     generate_jumps!(p, integrator, u, params, t)
     nothing
 end
@@ -79,7 +79,8 @@ function execute_jumps!(p::CCNRMJumpAggregation, integrator, u, params, t, affec
     u = update_state!(p, integrator, u, affects!)
 
     # update current jump rates and times
-    update_dependent_rates!(p, u, params, t)
+    rng = get_rng(integrator)
+    update_dependent_rates!(p, u, params, t, rng)
     nothing
 end
 
@@ -88,7 +89,7 @@ end
 function generate_jumps!(p::CCNRMJumpAggregation, integrator, u, params, t)
     p.next_jump, p.next_jump_time = getfirst(p.ptt)
 
-    # Rebuild the table if no next jump is found. 
+    # Rebuild the table if no next jump is found.
     if p.next_jump == 0
         timestep = 1 / sum(p.cur_rates)
         min_time = minimum(p.ptt.times)
@@ -102,7 +103,7 @@ end
 ######################## SSA specific helper routines ########################
 
 # Recalculate jump rates for jumps that depend on the just executed jump (p.next_jump)
-function update_dependent_rates!(p::CCNRMJumpAggregation, u, params, t)
+function update_dependent_rates!(p::CCNRMJumpAggregation, u, params, t, rng)
     @inbounds dep_rxs = p.dep_gr[p.next_jump]
     (; ptt, cur_rates, rates, ma_jumps, end_time) = p
     num_majumps = get_num_majumps(ma_jumps)
@@ -125,7 +126,7 @@ function update_dependent_rates!(p::CCNRMJumpAggregation, u, params, t)
             end
         else
             if cur_rates[rx] > zero(eltype(cur_rates))
-                update!(ptt, rx, oldtime, t + randexp(p.rng) / cur_rates[rx])
+                update!(ptt, rx, oldtime, t + randexp(rng) / cur_rates[rx])
             else
                 update!(ptt, rx, oldtime, floatmax(typeof(t)))
             end
@@ -134,15 +135,15 @@ function update_dependent_rates!(p::CCNRMJumpAggregation, u, params, t)
     nothing
 end
 
-# Evaluate all the rates and initialize the times in the priority table. 
-function initialize_rates_and_times!(p::CCNRMJumpAggregation, u, params, t)
+# Evaluate all the rates and initialize the times in the priority table.
+function initialize_rates_and_times!(p::CCNRMJumpAggregation, u, params, t, rng)
     # Initialize next-reaction times for the mass action jumps
     majumps = p.ma_jumps
     cur_rates = p.cur_rates
     pttdata = Vector{typeof(t)}(undef, length(cur_rates))
     @inbounds for i in 1:get_num_majumps(majumps)
         cur_rates[i] = evalrxrate(u, i, majumps)
-        pttdata[i] = t + randexp(p.rng) / cur_rates[i]
+        pttdata[i] = t + randexp(rng) / cur_rates[i]
     end
 
     # Initialize next-reaction times for the constant rates
@@ -150,11 +151,11 @@ function initialize_rates_and_times!(p::CCNRMJumpAggregation, u, params, t)
     idx = get_num_majumps(majumps) + 1
     @inbounds for rate in rates
         cur_rates[idx] = rate(u, params, t)
-        pttdata[idx] = t + randexp(p.rng) / cur_rates[idx]
+        pttdata[idx] = t + randexp(rng) / cur_rates[idx]
         idx += 1
     end
 
-    # Build the priority time table with the times and bin width. 
+    # Build the priority time table with the times and bin width.
     timestep = 1 / sum(cur_rates)
     p.ptt.times = pttdata
     rebuild!(p.ptt, t, timestep)