+++ src/aggregators/rssa.jl
@@ -4,8 +4,8 @@
 # functions of the current population sizes (i.e. u)
 # requires vartojumps_map and fluct_rates as JumpProblem keywords
 
-mutable struct RSSAJumpAggregation{T, S, F1, F2, RNG, VJMAP, JVMAP, BD, U} <:
-               AbstractSSAJumpAggregator{T, S, F1, F2, RNG}
+mutable struct RSSAJumpAggregation{T, S, F1, F2, VJMAP, JVMAP, BD, U} <:
+               AbstractSSAJumpAggregator{T, S, F1, F2}
     next_jump::Int
     prev_jump::Int
     next_jump_time::T
@@ -17,7 +17,6 @@ mutable struct RSSAJumpAggregation{T, S, F1, F2, RNG, VJMAP, JVMAP, BD, U} <:
     rates::F1
     affects!::F2
     save_positions::Tuple{Bool, Bool}
-    rng::RNG
     vartojumps_map::VJMAP
     jumptovars_map::JVMAP
     bracket_data::BD
@@ -26,10 +25,10 @@ mutable struct RSSAJumpAggregation{T, S, F1, F2, RNG, VJMAP, JVMAP, BD, U} <:
 end
 
 function RSSAJumpAggregation(nj::Int, njt::T, et::T, crs::Vector{T}, sr::T,
-        maj::S, rs::F1, affs!::F2, sps::Tuple{Bool, Bool},
-        rng::RNG; u::U, vartojumps_map = nothing,
+        maj::S, rs::F1, affs!::F2, sps::Tuple{Bool, Bool};
+        u::U, vartojumps_map = nothing,
         jumptovars_map = nothing,
-        bracket_data = nothing, kwargs...) where {T, S, F1, F2, RNG, U}
+        bracket_data = nothing, kwargs...) where {T, S, F1, F2, U}
     # a dependency graph is needed and must be provided if there are constant rate jumps
     if vartojumps_map === nothing
         if (get_num_majumps(maj) == 0) || !isempty(rs)
@@ -63,10 +62,10 @@ function RSSAJumpAggregation(nj::Int, njt::T, et::T, crs::Vector{T}, sr::T,
     uhigh = similar(u)
 
     affecttype = F2 <: Tuple ? F2 : Any
-    RSSAJumpAggregation{T, S, F1, affecttype, RNG, typeof(vtoj_map),
+    RSSAJumpAggregation{T, S, F1, affecttype, typeof(vtoj_map),
         typeof(jtov_map), typeof(bd), U}(nj, nj, njt, et, crl_bnds,
         crh_bnds, sr, maj, rs, affs!, sps,
-        rng, vtoj_map, jtov_map, bd, ulow,
+        vtoj_map, jtov_map, bd, ulow,
         uhigh)
 end
 
@@ -74,13 +73,13 @@ end
 
 # creating the JumpAggregation structure (function wrapper-based constant jumps)
 function aggregate(aggregator::RSSA, u, p, t, end_time, constant_jumps,
-        ma_jumps, save_positions, rng; kwargs...)
+        ma_jumps, save_positions; kwargs...)
 
     # handle constant jumps using function wrappers
     rates, affects! = get_jump_info_fwrappers(u, p, t, constant_jumps)
 
     build_jump_aggregation(RSSAJumpAggregation, u, p, t, end_time, ma_jumps,
-        rates, affects!, save_positions, rng; u = u,
+        rates, affects!, save_positions; u,
         kwargs...)
 end
 
@@ -108,7 +107,8 @@ function generate_jumps!(p::RSSAJumpAggregation, integrator, u, params, t)
         return nothing
     end
     # next jump type
-    (; ma_jumps, rates, cur_rate_high, cur_rate_low, rng) = p
+    (; ma_jumps, rates, cur_rate_high, cur_rate_low) = p
+    rng = get_rng(integrator)
     num_majumps = get_num_majumps(ma_jumps)
     rerl = zero(sum_rate)