+++ src/aggregators/rssacr.jl
@@ -4,9 +4,9 @@ Composition-Rejection with Rejection sampling method (RSSA-CR)
 
 const MINJUMPRATE = 2.0^exponent(1e-12)
 
-mutable struct RSSACRJumpAggregation{F, S, F1, F2, RNG, U, VJMAP, JVMAP, BD,
+mutable struct RSSACRJumpAggregation{F, S, F1, F2, U, VJMAP, JVMAP, BD,
     P <: PriorityTable, W <: Function} <:
-               AbstractSSAJumpAggregator{F, S, F1, F2, RNG}
+               AbstractSSAJumpAggregator{F, S, F1, F2}
     next_jump::Int
     prev_jump::Int
     next_jump_time::F
@@ -18,7 +18,6 @@ mutable struct RSSACRJumpAggregation{F, S, F1, F2, RNG, U, VJMAP, JVMAP, BD,
     rates::F1
     affects!::F2
     save_positions::Tuple{Bool, Bool}
-    rng::RNG
     vartojumps_map::VJMAP
     jumptovars_map::JVMAP
     bracket_data::BD
@@ -31,11 +30,11 @@ mutable struct RSSACRJumpAggregation{F, S, F1, F2, RNG, U, VJMAP, JVMAP, BD,
 end
 
 function RSSACRJumpAggregation(nj::Int, njt::F, et::F, crs::Vector{F}, sum_rate::F, maj::S,
-        rs::F1, affs!::F2, sps::Tuple{Bool, Bool}, rng::RNG; u::U,
+        rs::F1, affs!::F2, sps::Tuple{Bool, Bool}; u::U,
         vartojumps_map = nothing, jumptovars_map = nothing,
         bracket_data = nothing, minrate = convert(F, MINJUMPRATE),
         maxrate = convert(F, Inf),
-        kwargs...) where {F, S, F1, F2, RNG, U}
+        kwargs...) where {F, S, F1, F2, U}
     # a dependency graph is needed and must be provided if there are constant rate jumps
     if vartojumps_map === nothing
         if (get_num_majumps(maj) == 0) || !isempty(rs)
@@ -80,10 +79,10 @@ function RSSACRJumpAggregation(nj::Int, njt::F, et::F, crs::Vector{F}, sum_rate:
     rt = PriorityTable(ratetogroup, zeros(F, 1), minrate, 2 * minrate)
 
     affecttype = F2 <: Tuple ? F2 : Any
-    RSSACRJumpAggregation{typeof(njt), S, F1, affecttype, RNG, U, typeof(vtoj_map),
+    RSSACRJumpAggregation{typeof(njt), S, F1, affecttype, U, typeof(vtoj_map),
         typeof(jtov_map), typeof(bd), typeof(rt),
         typeof(ratetogroup)}(nj, nj, njt, et, crl_bnds, crh_bnds,
-        sum_rate, maj, rs, affs!, sps, rng, vtoj_map,
+        sum_rate, maj, rs, affs!, sps, vtoj_map,
         jtov_map, bd, ulow, uhigh, minrate, maxrate,
         rt, ratetogroup)
 end
@@ -92,13 +91,13 @@ end
 
 # creating the JumpAggregation structure (function wrapper-based constant jumps)
 function aggregate(aggregator::RSSACR, u, p, t, end_time, constant_jumps,
-        ma_jumps, save_positions, rng; kwargs...)
+        ma_jumps, save_positions; kwargs...)
 
     # handle constant jumps using function wrappers
     rates, affects! = get_jump_info_fwrappers(u, p, t, constant_jumps)
 
     build_jump_aggregation(RSSACRJumpAggregation, u, p, t, end_time, ma_jumps,
-        rates, affects!, save_positions, rng; u = u, kwargs...)
+        rates, affects!, save_positions; u, kwargs...)
 end
 
 # set up a new simulation and calculate the first jump / jump time
@@ -134,7 +133,8 @@ function generate_jumps!(p::RSSACRJumpAggregation, integrator, u, params, t)
         return nothing
     end
 
-    (; rt, ma_jumps, rates, cur_rate_high, cur_rate_low, rng) = p
+    (; rt, ma_jumps, rates, cur_rate_high, cur_rate_low) = p
+    rng = get_rng(integrator)
     num_majumps = get_num_majumps(ma_jumps)
     rerl = zero(sum_rate)